Refactor error handling into its own module
